+++ tests/rest/create_db.sh
@@ -17,4 +17,5 @@ EOF
 export PGPASSWORD='password'
 psql -h localhost -d test_dispatcher -U tester -a -f tests/rest/dispatcher.sql
 psql -h localhost -d test_eva -U tester -a -f tests/rest/eva.sql
+psql -h localhost -d test_eva -U tester -a -f tests/rest/eva_quiz.sql