+++ Changelog.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.5] - 2020-05-19
 ### Changed
-- Fixed troubles with db connector for jobs handlers.
+- Fixed troubles with pool in db connector for jobs handlers.
 
 ## [1.0.4] - 2020-05-07
 ### Changed