+++ Changelog.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Quizs features
+
 ## [1.0.4] - 2020-05-07
 ### Changed
 - Fixed using cache\_ttl param from spl query.